refactor: split transpiled code
