+++ .gitattributes
@@ -1,1 +1,1 @@
-lua/typescript.lua linguist-generated=true
+lua/typescript/** linguist-generated
