+++ lua/typescript/init.lua
@@ -1,3021 +1,22 @@
 --[[ Generated with https://github.com/TypeScriptToLua/TypeScriptToLua ]]
-
-local ____modules = {}
-local ____moduleCache = {}
-local ____originalRequire = require
-local function require(file, ...)
-    if ____moduleCache[file] then
-        return ____moduleCache[file].value
-    end
-    if ____modules[file] then
-        local module = ____modules[file]
-        ____moduleCache[file] = { value = (select("#", ...) > 0) and module(...) or module(file) }
-        return ____moduleCache[file].value
-    else
-        if ____originalRequire then
-            return ____originalRequire(file)
-        else
-            error("module '" .. file .. "' not found")
-        end
-    end
-end
-____modules = {
-["lualib_bundle"] = function(...) 
-local function __TS__ArrayIsArray(value)
-    return type(value) == "table" and (value[1] ~= nil or next(value) == nil)
-end
-
-local function __TS__ArrayConcat(self, ...)
-    local items = {...}
-    local result = {}
-    local len = 0
-    for i = 1, #self do
-        len = len + 1
-        result[len] = self[i]
-    end
-    for i = 1, #items do
-        local item = items[i]
-        if __TS__ArrayIsArray(item) then
-            for j = 1, #item do
-                len = len + 1
-                result[len] = item[j]
-            end
-        else
-            len = len + 1
-            result[len] = item
-        end
-    end
-    return result
-end
-
-local __TS__Symbol, Symbol
-do
-    local symbolMetatable = {__tostring = function(self)
-        return ("Symbol(" .. (self.description or "")) .. ")"
-    end}
-    function __TS__Symbol(description)
-        return setmetatable({description = description}, symbolMetatable)
-    end
-    Symbol = {
-        iterator = __TS__Symbol("Symbol.iterator"),
-        hasInstance = __TS__Symbol("Symbol.hasInstance"),
-        species = __TS__Symbol("Symbol.species"),
-        toStringTag = __TS__Symbol("Symbol.toStringTag")
-    }
-end
-
-local function __TS__ArrayEntries(array)
-    local key = 0
-    return {
-        [Symbol.iterator] = function(self)
-            return self
-        end,
-        next = function(self)
-            local result = {done = array[key + 1] == nil, value = {key, array[key + 1]}}
-            key = key + 1
-            return result
-        end
-    }
-end
-
-local function __TS__ArrayEvery(self, callbackfn, thisArg)
-    for i = 1, #self do
-        if not callbackfn(thisArg, self[i], i - 1, self) then
-            return false
-        end
-    end
-    return true
-end
-
-local function __TS__ArrayFilter(self, callbackfn, thisArg)
-    local result = {}
-    local len = 0
-    for i = 1, #self do
-        if callbackfn(thisArg, self[i], i - 1, self) then
-            len = len + 1
-            result[len] = self[i]
-        end
-    end
-    return result
-end
-
-local function __TS__ArrayForEach(self, callbackFn, thisArg)
-    for i = 1, #self do
-        callbackFn(thisArg, self[i], i - 1, self)
-    end
-end
-
-local function __TS__ArrayFind(self, predicate, thisArg)
-    for i = 1, #self do
-        local elem = self[i]
-        if predicate(thisArg, elem, i - 1, self) then
-            return elem
-        end
-    end
-    return nil
-end
-
-local function __TS__ArrayFindIndex(self, callbackFn, thisArg)
-    for i = 1, #self do
-        if callbackFn(thisArg, self[i], i - 1, self) then
-            return i - 1
-        end
-    end
-    return -1
-end
-
-local __TS__Iterator
-do
-    local function iteratorGeneratorStep(self)
-        local co = self.____coroutine
-        local status, value = coroutine.resume(co)
-        if not status then
-            error(value, 0)
-        end
-        if coroutine.status(co) == "dead" then
-            return
-        end
-        return true, value
-    end
-    local function iteratorIteratorStep(self)
-        local result = self:next()
-        if result.done then
-            return
-        end
-        return true, result.value
-    end
-    local function iteratorStringStep(self, index)
-        index = index + 1
-        if index > #self then
-            return
-        end
-        return index, string.sub(self, index, index)
-    end
-    function __TS__Iterator(iterable)
-        if type(iterable) == "string" then
-            return iteratorStringStep, iterable, 0
-        elseif iterable.____coroutine ~= nil then
-            return iteratorGeneratorStep, iterable
-        elseif iterable[Symbol.iterator] then
-            local iterator = iterable[Symbol.iterator](iterable)
-            return iteratorIteratorStep, iterator
-        else
-            return ipairs(iterable)
-        end
-    end
-end
-
-local __TS__ArrayFrom
-do
-    local function arrayLikeStep(self, index)
-        index = index + 1
-        if index > self.length then
-            return
-        end
-        return index, self[index]
-    end
-    local function arrayLikeIterator(arr)
-        if type(arr.length) == "number" then
-            return arrayLikeStep, arr, 0
-        end
-        return __TS__Iterator(arr)
-    end
-    function __TS__ArrayFrom(arrayLike, mapFn, thisArg)
-        local result = {}
-        if mapFn == nil then
-            for ____, v in arrayLikeIterator(arrayLike) do
-                result[#result + 1] = v
-            end
-        else
-            for i, v in arrayLikeIterator(arrayLike) do
-                result[#result + 1] = mapFn(thisArg, v, i - 1)
-            end
-        end
-        return result
-    end
-end
-
-local function __TS__ArrayIncludes(self, searchElement, fromIndex)
-    if fromIndex == nil then
-        fromIndex = 0
-    end
-    local len = #self
-    local k = fromIndex
-    if fromIndex < 0 then
-        k = len + fromIndex
-    end
-    if k < 0 then
-        k = 0
-    end
-    for i = k + 1, len do
-        if self[i] == searchElement then
-            return true
-        end
-    end
-    return false
-end
-
-local function __TS__ArrayIndexOf(self, searchElement, fromIndex)
-    if fromIndex == nil then
-        fromIndex = 0
-    end
-    local len = #self
-    if len == 0 then
-        return -1
-    end
-    if fromIndex >= len then
-        return -1
-    end
-    if fromIndex < 0 then
-        fromIndex = len + fromIndex
-        if fromIndex < 0 then
-            fromIndex = 0
-        end
-    end
-    for i = fromIndex + 1, len do
-        if self[i] == searchElement then
-            return i - 1
-        end
-    end
-    return -1
-end
-
-local function __TS__ArrayJoin(self, separator)
-    if separator == nil then
-        separator = ","
-    end
-    local parts = {}
-    for i = 1, #self do
-        parts[i] = tostring(self[i])
-    end
-    return table.concat(parts, separator)
-end
-
-local function __TS__ArrayMap(self, callbackfn, thisArg)
-    local result = {}
-    for i = 1, #self do
-        result[i] = callbackfn(thisArg, self[i], i - 1, self)
-    end
-    return result
-end
-
-local function __TS__ArrayPush(self, ...)
-    local items = {...}
-    local len = #self
-    for i = 1, #items do
-        len = len + 1
-        self[len] = items[i]
-    end
-    return len
-end
-
-local function __TS__ArrayPushArray(self, items)
-    local len = #self
-    for i = 1, #items do
-        len = len + 1
-        self[len] = items[i]
-    end
-    return len
-end
-
-local function __TS__CountVarargs(...)
-    return select("#", ...)
-end
-
-local function __TS__ArrayReduce(self, callbackFn, ...)
-    local len = #self
-    local k = 0
-    local accumulator = nil
-    if __TS__CountVarargs(...) ~= 0 then
-        accumulator = ...
-    elseif len > 0 then
-        accumulator = self[1]
-        k = 1
-    else
-        error("Reduce of empty array with no initial value", 0)
-    end
-    for i = k + 1, len do
-        accumulator = callbackFn(
-            nil,
-            accumulator,
-            self[i],
-            i - 1,
-            self
-        )
-    end
-    return accumulator
-end
-
-local function __TS__ArrayReduceRight(self, callbackFn, ...)
-    local len = #self
-    local k = len - 1
-    local accumulator = nil
-    if __TS__CountVarargs(...) ~= 0 then
-        accumulator = ...
-    elseif len > 0 then
-        accumulator = self[k + 1]
-        k = k - 1
-    else
-        error("Reduce of empty array with no initial value", 0)
-    end
-    for i = k + 1, 1, -1 do
-        accumulator = callbackFn(
-            nil,
-            accumulator,
-            self[i],
-            i - 1,
-            self
-        )
-    end
-    return accumulator
-end
-
-local function __TS__ArrayReverse(self)
-    local i = 1
-    local j = #self
-    while i < j do
-        local temp = self[j]
-        self[j] = self[i]
-        self[i] = temp
-        i = i + 1
-        j = j - 1
-    end
-    return self
-end
-
-local function __TS__ArrayUnshift(self, ...)
-    local items = {...}
-    local numItemsToInsert = #items
-    if numItemsToInsert == 0 then
-        return #self
-    end
-    for i = #self, 1, -1 do
-        self[i + numItemsToInsert] = self[i]
-    end
-    for i = 1, numItemsToInsert do
-        self[i] = items[i]
-    end
-    return #self
-end
-
-local function __TS__ArraySort(self, compareFn)
-    if compareFn ~= nil then
-        table.sort(
-            self,
-            function(a, b) return compareFn(nil, a, b) < 0 end
-        )
-    else
-        table.sort(self)
-    end
-    return self
-end
-
-local function __TS__ArraySlice(self, first, last)
-    local len = #self
-    local ____first_0 = first
-    if ____first_0 == nil then
-        ____first_0 = 0
-    end
-    first = ____first_0
-    if first < 0 then
-        first = len + first
-        if first < 0 then
-            first = 0
-        end
-    else
-        if first > len then
-            first = len
-        end
-    end
-    local ____last_1 = last
-    if ____last_1 == nil then
-        ____last_1 = len
-    end
-    last = ____last_1
-    if last < 0 then
-        last = len + last
-        if last < 0 then
-            last = 0
-        end
-    else
-        if last > len then
-            last = len
-        end
-    end
-    local out = {}
-    first = first + 1
-    last = last + 1
-    local n = 1
-    while first < last do
-        out[n] = self[first]
-        first = first + 1
-        n = n + 1
-    end
-    return out
-end
-
-local function __TS__ArraySome(self, callbackfn, thisArg)
-    for i = 1, #self do
-        if callbackfn(thisArg, self[i], i - 1, self) then
-            return true
-        end
-    end
-    return false
-end
-
-local function __TS__ArraySplice(self, ...)
-    local args = {...}
-    local len = #self
-    local actualArgumentCount = __TS__CountVarargs(...)
-    local start = args[1]
-    local deleteCount = args[2]
-    if start < 0 then
-        start = len + start
-        if start < 0 then
-            start = 0
-        end
-    elseif start > len then
-        start = len
-    end
-    local itemCount = actualArgumentCount - 2
-    if itemCount < 0 then
-        itemCount = 0
-    end
-    local actualDeleteCount
-    if actualArgumentCount == 0 then
-        actualDeleteCount = 0
-    elseif actualArgumentCount == 1 then
-        actualDeleteCount = len - start
-    else
-        local ____deleteCount_0 = deleteCount
-        if ____deleteCount_0 == nil then
-            ____deleteCount_0 = 0
-        end
-        actualDeleteCount = ____deleteCount_0
-        if actualDeleteCount < 0 then
-            actualDeleteCount = 0
-        end
-        if actualDeleteCount > len - start then
-            actualDeleteCount = len - start
-        end
-    end
-    local out = {}
-    for k = 1, actualDeleteCount do
-        local from = start + k
-        if self[from] ~= nil then
-            out[k] = self[from]
-        end
-    end
-    if itemCount < actualDeleteCount then
-        for k = start + 1, len - actualDeleteCount do
-            local from = k + actualDeleteCount
-            local to = k + itemCount
-            if self[from] then
-                self[to] = self[from]
-            else
-                self[to] = nil
-            end
-        end
-        for k = len - actualDeleteCount + itemCount + 1, len do
-            self[k] = nil
-        end
-    elseif itemCount > actualDeleteCount then
-        for k = len - actualDeleteCount, start + 1, -1 do
-            local from = k + actualDeleteCount
-            local to = k + itemCount
-            if self[from] then
-                self[to] = self[from]
-            else
-                self[to] = nil
-            end
-        end
-    end
-    local j = start + 1
-    for i = 3, actualArgumentCount do
-        self[j] = args[i]
-        j = j + 1
-    end
-    for k = #self, len - actualDeleteCount + itemCount + 1, -1 do
-        self[k] = nil
-    end
-    return out
-end
-
-local function __TS__ArrayToObject(self)
-    local object = {}
-    for i = 1, #self do
-        object[i - 1] = self[i]
-    end
-    return object
-end
-
-local function __TS__ArrayFlat(self, depth)
-    if depth == nil then
-        depth = 1
-    end
-    local result = {}
-    local len = 0
-    for i = 1, #self do
-        local value = self[i]
-        if depth > 0 and __TS__ArrayIsArray(value) then
-            local toAdd
-            if depth == 1 then
-                toAdd = value
-            else
-                toAdd = __TS__ArrayFlat(value, depth - 1)
-            end
-            for j = 1, #toAdd do
-                local val = toAdd[j]
-                len = len + 1
-                result[len] = val
-            end
-        else
-            len = len + 1
-            result[len] = value
-        end
-    end
-    return result
-end
-
-local function __TS__ArrayFlatMap(self, callback, thisArg)
-    local result = {}
-    local len = 0
-    for i = 1, #self do
-        local value = callback(thisArg, self[i], i - 1, self)
-        if __TS__ArrayIsArray(value) then
-            for j = 1, #value do
-                len = len + 1
-                result[len] = value[j]
-            end
-        else
-            len = len + 1
-            result[len] = value
-        end
-    end
-    return result
-end
-
-local function __TS__ArraySetLength(self, length)
-    if length < 0 or length ~= length or length == math.huge or math.floor(length) ~= length then
-        error(
-            "invalid array length: " .. tostring(length),
-            0
-        )
-    end
-    for i = length + 1, #self do
-        self[i] = nil
-    end
-    return length
-end
-
-local function __TS__InstanceOf(obj, classTbl)
-    if type(classTbl) ~= "table" then
-        error("Right-hand side of 'instanceof' is not an object", 0)
-    end
-    if classTbl[Symbol.hasInstance] ~= nil then
-        return not not classTbl[Symbol.hasInstance](classTbl, obj)
-    end
-    if type(obj) == "table" then
-        local luaClass = obj.constructor
-        while luaClass ~= nil do
-            if luaClass == classTbl then
-                return true
-            end
-            luaClass = luaClass.____super
-        end
-    end
-    return false
-end
-
-local function __TS__New(target, ...)
-    local instance = setmetatable({}, target.prototype)
-    instance:____constructor(...)
-    return instance
-end
-
-local function __TS__Class(self)
-    local c = {prototype = {}}
-    c.prototype.__index = c.prototype
-    c.prototype.constructor = c
-    return c
-end
-
-local __TS__Unpack = table.unpack or unpack
-
-local function __TS__FunctionBind(fn, ...)
-    local boundArgs = {...}
-    return function(____, ...)
-        local args = {...}
-        __TS__ArrayUnshift(
-            args,
-            __TS__Unpack(boundArgs)
-        )
-        return fn(__TS__Unpack(args))
-    end
-end
-
-local __TS__Promise
-do
-    local function promiseDeferred(self)
-        local resolve
-        local reject
-        local promise = __TS__New(
-            __TS__Promise,
-            function(____, res, rej)
-                resolve = res
-                reject = rej
-            end
-        )
-        return {promise = promise, resolve = resolve, reject = reject}
-    end
-    local function isPromiseLike(self, thing)
-        return __TS__InstanceOf(thing, __TS__Promise)
-    end
-    __TS__Promise = __TS__Class()
-    __TS__Promise.name = "__TS__Promise"
-    function __TS__Promise.prototype.____constructor(self, executor)
-        self.state = 0
-        self.fulfilledCallbacks = {}
-        self.rejectedCallbacks = {}
-        self.finallyCallbacks = {}
-        do
-            local function ____catch(e)
-                self:reject(e)
-            end
-            local ____try, ____hasReturned = pcall(function()
-                executor(
-                    nil,
-                    __TS__FunctionBind(self.resolve, self),
-                    __TS__FunctionBind(self.reject, self)
-                )
-            end)
-            if not ____try then
-                ____catch(____hasReturned)
-            end
-        end
-    end
-    function __TS__Promise.resolve(data)
-        local promise = __TS__New(
-            __TS__Promise,
-            function()
-            end
-        )
-        promise.state = 1
-        promise.value = data
-        return promise
-    end
-    function __TS__Promise.reject(reason)
-        local promise = __TS__New(
-            __TS__Promise,
-            function()
-            end
-        )
-        promise.state = 2
-        promise.rejectionReason = reason
-        return promise
-    end
-    __TS__Promise.prototype["then"] = function(self, onFulfilled, onRejected)
-        local ____promiseDeferred_result_0 = promiseDeferred(nil)
-        local promise = ____promiseDeferred_result_0.promise
-        local resolve = ____promiseDeferred_result_0.resolve
-        local reject = ____promiseDeferred_result_0.reject
-        local isFulfilled = self.state == 1
-        local isRejected = self.state == 2
-        if onFulfilled then
-            local internalCallback = self:createPromiseResolvingCallback(onFulfilled, resolve, reject)
-            local ____self_fulfilledCallbacks_1 = self.fulfilledCallbacks
-            ____self_fulfilledCallbacks_1[#____self_fulfilledCallbacks_1 + 1] = internalCallback
-            if isFulfilled then
-                internalCallback(nil, self.value)
-            end
-        else
-            local ____self_fulfilledCallbacks_2 = self.fulfilledCallbacks
-            ____self_fulfilledCallbacks_2[#____self_fulfilledCallbacks_2 + 1] = function(____, v) return resolve(nil, v) end
-        end
-        if onRejected then
-            local internalCallback = self:createPromiseResolvingCallback(onRejected, resolve, reject)
-            local ____self_rejectedCallbacks_3 = self.rejectedCallbacks
-            ____self_rejectedCallbacks_3[#____self_rejectedCallbacks_3 + 1] = internalCallback
-            if isRejected then
-                internalCallback(nil, self.rejectionReason)
-            end
-        else
-            local ____self_rejectedCallbacks_4 = self.rejectedCallbacks
-            ____self_rejectedCallbacks_4[#____self_rejectedCallbacks_4 + 1] = function(____, err) return reject(nil, err) end
-        end
-        if isFulfilled then
-            resolve(nil, self.value)
-        end
-        if isRejected then
-            reject(nil, self.rejectionReason)
-        end
-        return promise
-    end
-    function __TS__Promise.prototype.catch(self, onRejected)
-        return self["then"](self, nil, onRejected)
-    end
-    function __TS__Promise.prototype.finally(self, onFinally)
-        if onFinally then
-            local ____self_finallyCallbacks_5 = self.finallyCallbacks
-            ____self_finallyCallbacks_5[#____self_finallyCallbacks_5 + 1] = onFinally
-            if self.state ~= 0 then
-                onFinally(nil)
-            end
-        end
-        return self
-    end
-    function __TS__Promise.prototype.resolve(self, data)
-        if __TS__InstanceOf(data, __TS__Promise) then
-            data["then"](
-                data,
-                function(____, v) return self:resolve(v) end,
-                function(____, err) return self:reject(err) end
-            )
-            return
-        end
-        if self.state == 0 then
-            self.state = 1
-            self.value = data
-            for ____, callback in ipairs(self.fulfilledCallbacks) do
-                callback(nil, data)
-            end
-            for ____, callback in ipairs(self.finallyCallbacks) do
-                callback(nil)
-            end
-        end
-    end
-    function __TS__Promise.prototype.reject(self, reason)
-        if self.state == 0 then
-            self.state = 2
-            self.rejectionReason = reason
-            for ____, callback in ipairs(self.rejectedCallbacks) do
-                callback(nil, reason)
-            end
-            for ____, callback in ipairs(self.finallyCallbacks) do
-                callback(nil)
-            end
-        end
-    end
-    function __TS__Promise.prototype.createPromiseResolvingCallback(self, f, resolve, reject)
-        return function(____, value)
-            do
-                local function ____catch(e)
-                    reject(nil, e)
-                end
-                local ____try, ____hasReturned = pcall(function()
-                    self:handleCallbackData(
-                        f(nil, value),
-                        resolve,
-                        reject
-                    )
-                end)
-                if not ____try then
-                    ____catch(____hasReturned)
-                end
-            end
-        end
-    end
-    function __TS__Promise.prototype.handleCallbackData(self, data, resolve, reject)
-        if isPromiseLike(nil, data) then
-            local nextpromise = data
-            if nextpromise.state == 1 then
-                resolve(nil, nextpromise.value)
-            elseif nextpromise.state == 2 then
-                reject(nil, nextpromise.rejectionReason)
-            else
-                data["then"](data, resolve, reject)
-            end
-        else
-            resolve(nil, data)
-        end
-    end
-end
-
-local function __TS__AsyncAwaiter(generator)
-    return __TS__New(
-        __TS__Promise,
-        function(____, resolve, reject)
-            local adopt, fulfilled, step, resolved, asyncCoroutine
-            function adopt(self, value)
-                local ____temp_0
-                if __TS__InstanceOf(value, __TS__Promise) then
-                    ____temp_0 = value
-                else
-                    ____temp_0 = __TS__Promise.resolve(value)
-                end
-                return ____temp_0
-            end
-            function fulfilled(self, value)
-                local success, resultOrError = coroutine.resume(asyncCoroutine, value)
-                if success then
-                    step(nil, resultOrError)
-                else
-                    reject(nil, resultOrError)
-                end
-            end
-            function step(self, result)
-                if resolved then
-                    return
-                end
-                if coroutine.status(asyncCoroutine) == "dead" then
-                    resolve(nil, result)
-                else
-                    local ____self_1 = adopt(nil, result)
-                    ____self_1["then"](____self_1, fulfilled, reject)
-                end
-            end
-            resolved = false
-            asyncCoroutine = coroutine.create(generator)
-            local success, resultOrError = coroutine.resume(
-                asyncCoroutine,
-                function(____, v)
-                    resolved = true
-                    local ____self_2 = adopt(nil, v)
-                    ____self_2["then"](____self_2, resolve, reject)
-                end
-            )
-            if success then
-                step(nil, resultOrError)
-            else
-                reject(nil, resultOrError)
-            end
-        end
-    )
-end
-local function __TS__Await(thing)
-    return coroutine.yield(thing)
-end
-
-local function __TS__ClassExtends(target, base)
-    target.____super = base
-    local staticMetatable = setmetatable({__index = base}, base)
-    setmetatable(target, staticMetatable)
-    local baseMetatable = getmetatable(base)
-    if baseMetatable then
-        if type(baseMetatable.__index) == "function" then
-            staticMetatable.__index = baseMetatable.__index
-        end
-        if type(baseMetatable.__newindex) == "function" then
-            staticMetatable.__newindex = baseMetatable.__newindex
-        end
-    end
-    setmetatable(target.prototype, base.prototype)
-    if type(base.prototype.__index) == "function" then
-        target.prototype.__index = base.prototype.__index
-    end
-    if type(base.prototype.__newindex) == "function" then
-        target.prototype.__newindex = base.prototype.__newindex
-    end
-    if type(base.prototype.__tostring) == "function" then
-        target.prototype.__tostring = base.prototype.__tostring
-    end
-end
-
-local function __TS__CloneDescriptor(____bindingPattern0)
-    local value
-    local writable
-    local set
-    local get
-    local configurable
-    local enumerable
-    enumerable = ____bindingPattern0.enumerable
-    configurable = ____bindingPattern0.configurable
-    get = ____bindingPattern0.get
-    set = ____bindingPattern0.set
-    writable = ____bindingPattern0.writable
-    value = ____bindingPattern0.value
-    local descriptor = {enumerable = enumerable == true, configurable = configurable == true}
-    local hasGetterOrSetter = get ~= nil or set ~= nil
-    local hasValueOrWritableAttribute = writable ~= nil or value ~= nil
-    if hasGetterOrSetter and hasValueOrWritableAttribute then
-        error("Invalid property descriptor. Cannot both specify accessors and a value or writable attribute.", 0)
-    end
-    if get or set then
-        descriptor.get = get
-        descriptor.set = set
-    else
-        descriptor.value = value
-        descriptor.writable = writable == true
-    end
-    return descriptor
-end
-
-local function __TS__ObjectAssign(target, ...)
-    local sources = {...}
-    for i = 1, #sources do
-        local source = sources[i]
-        for key in pairs(source) do
-            target[key] = source[key]
-        end
-    end
-    return target
-end
-
-local function __TS__ObjectGetOwnPropertyDescriptor(object, key)
-    local metatable = getmetatable(object)
-    if not metatable then
-        return
-    end
-    if not rawget(metatable, "_descriptors") then
-        return
-    end
-    return rawget(metatable, "_descriptors")[key]
-end
-
-local __TS__SetDescriptor
-do
-    local function descriptorIndex(self, key)
-        local value = rawget(self, key)
-        if value ~= nil then
-            return value
-        end
-        local metatable = getmetatable(self)
-        while metatable do
-            local rawResult = rawget(metatable, key)
-            if rawResult ~= nil then
-                return rawResult
-            end
-            local descriptors = rawget(metatable, "_descriptors")
-            if descriptors then
-                local descriptor = descriptors[key]
-                if descriptor then
-                    if descriptor.get then
-                        return descriptor.get(self)
-                    end
-                    return descriptor.value
-                end
-            end
-            metatable = getmetatable(metatable)
-        end
-    end
-    local function descriptorNewIndex(self, key, value)
-        local metatable = getmetatable(self)
-        while metatable do
-            local descriptors = rawget(metatable, "_descriptors")
-            if descriptors then
-                local descriptor = descriptors[key]
-                if descriptor then
-                    if descriptor.set then
-                        descriptor.set(self, value)
-                    else
-                        if descriptor.writable == false then
-                            error(
-                                ((("Cannot assign to read only property '" .. key) .. "' of object '") .. tostring(self)) .. "'",
-                                0
-                            )
-                        end
-                        descriptor.value = value
-                    end
-                    return
-                end
-            end
-            metatable = getmetatable(metatable)
-        end
-        rawset(self, key, value)
-    end
-    function __TS__SetDescriptor(target, key, desc, isPrototype)
-        if isPrototype == nil then
-            isPrototype = false
-        end
-        local ____isPrototype_0
-        if isPrototype then
-            ____isPrototype_0 = target
-        else
-            ____isPrototype_0 = getmetatable(target)
-        end
-        local metatable = ____isPrototype_0
-        if not metatable then
-            metatable = {}
-            setmetatable(target, metatable)
-        end
-        local value = rawget(target, key)
-        if value ~= nil then
-            rawset(target, key, nil)
-        end
-        if not rawget(metatable, "_descriptors") then
-            metatable._descriptors = {}
-        end
-        local descriptor = __TS__CloneDescriptor(desc)
-        metatable._descriptors[key] = descriptor
-        metatable.__index = descriptorIndex
-        metatable.__newindex = descriptorNewIndex
-    end
-end
-
-local function __TS__Decorate(decorators, target, key, desc)
-    local result = target
-    do
-        local i = #decorators
-        while i >= 0 do
-            local decorator = decorators[i + 1]
-            if decorator then
-                local oldResult = result
-                if key == nil then
-                    result = decorator(nil, result)
-                elseif desc == true then
-                    local value = rawget(target, key)
-                    local descriptor = __TS__ObjectGetOwnPropertyDescriptor(target, key) or ({configurable = true, writable = true, value = value})
-                    local desc = decorator(nil, target, key, descriptor) or descriptor
-                    local isSimpleValue = desc.configurable == true and desc.writable == true and not desc.get and not desc.set
-                    if isSimpleValue then
-                        rawset(target, key, desc.value)
-                    else
-                        __TS__SetDescriptor(
-                            target,
-                            key,
-                            __TS__ObjectAssign({}, descriptor, desc)
-                        )
-                    end
-                elseif desc == false then
-                    result = decorator(nil, target, key, desc)
-                else
-                    result = decorator(nil, target, key)
-                end
-                result = result or oldResult
-            end
-            i = i - 1
-        end
-    end
-    return result
-end
-
-local function __TS__DecorateParam(paramIndex, decorator)
-    return function(____, target, key) return decorator(nil, target, key, paramIndex) end
-end
-
-local function __TS__StringIncludes(self, searchString, position)
-    if not position then
-        position = 1
-    else
-        position = position + 1
-    end
-    local index = string.find(self, searchString, position, true)
-    return index ~= nil
-end
-
-local Error, RangeError, ReferenceError, SyntaxError, TypeError, URIError
-do
-    local function getErrorStack(self, constructor)
-        local level = 1
-        while true do
-            local info = debug.getinfo(level, "f")
-            level = level + 1
-            if not info then
-                level = 1
-                break
-            elseif info.func == constructor then
-                break
-            end
-        end
-        if __TS__StringIncludes(_VERSION, "Lua 5.0") then
-            return debug.traceback(("[Level " .. tostring(level)) .. "]")
-        else
-            return debug.traceback(nil, level)
-        end
-    end
-    local function wrapErrorToString(self, getDescription)
-        return function(self)
-            local description = getDescription(self)
-            local caller = debug.getinfo(3, "f")
-            local isClassicLua = __TS__StringIncludes(_VERSION, "Lua 5.0") or _VERSION == "Lua 5.1"
-            if isClassicLua or caller and caller.func ~= error then
-                return description
-            else
-                return (tostring(description) .. "\n") .. self.stack
-            end
-        end
-    end
-    local function initErrorClass(self, Type, name)
-        Type.name = name
-        return setmetatable(
-            Type,
-            {__call = function(____, _self, message) return __TS__New(Type, message) end}
-        )
-    end
-    local ____initErrorClass_2 = initErrorClass
-    local ____class_0 = __TS__Class()
-    ____class_0.name = ""
-    function ____class_0.prototype.____constructor(self, message)
-        if message == nil then
-            message = ""
-        end
-        self.message = message
-        self.name = "Error"
-        self.stack = getErrorStack(nil, self.constructor.new)
-        local metatable = getmetatable(self)
-        if not metatable.__errorToStringPatched then
-            metatable.__errorToStringPatched = true
-            metatable.__tostring = wrapErrorToString(nil, metatable.__tostring)
-        end
-    end
-    function ____class_0.prototype.__tostring(self)
-        local ____temp_1
-        if self.message ~= "" then
-            ____temp_1 = (self.name .. ": ") .. self.message
-        else
-            ____temp_1 = self.name
-        end
-        return ____temp_1
-    end
-    Error = ____initErrorClass_2(nil, ____class_0, "Error")
-    local function createErrorClass(self, name)
-        local ____initErrorClass_4 = initErrorClass
-        local ____class_3 = __TS__Class()
-        ____class_3.name = ____class_3.name
-        __TS__ClassExtends(____class_3, Error)
-        function ____class_3.prototype.____constructor(self, ...)
-            ____class_3.____super.prototype.____constructor(self, ...)
-            self.name = name
-        end
-        return ____initErrorClass_4(nil, ____class_3, name)
-    end
-    RangeError = createErrorClass(nil, "RangeError")
-    ReferenceError = createErrorClass(nil, "ReferenceError")
-    SyntaxError = createErrorClass(nil, "SyntaxError")
-    TypeError = createErrorClass(nil, "TypeError")
-    URIError = createErrorClass(nil, "URIError")
-end
-
-local function __TS__ObjectGetOwnPropertyDescriptors(object)
-    local metatable = getmetatable(object)
-    if not metatable then
-        return {}
-    end
-    return rawget(metatable, "_descriptors") or ({})
-end
-
-local function __TS__Delete(target, key)
-    local descriptors = __TS__ObjectGetOwnPropertyDescriptors(target)
-    local descriptor = descriptors[key]
-    if descriptor then
-        if not descriptor.configurable then
-            error(
-                __TS__New(
-                    TypeError,
-                    ((("Cannot delete property " .. tostring(key)) .. " of ") .. tostring(target)) .. "."
-                ),
-                0
-            )
-        end
-        descriptors[key] = nil
-        return true
-    end
-    target[key] = nil
-    return true
-end
-
-local function __TS__StringAccess(self, index)
-    if index >= 0 and index < #self then
-        return string.sub(self, index + 1, index + 1)
-    end
-end
-
-local function __TS__DelegatedYield(iterable)
-    if type(iterable) == "string" then
-        for index = 0, #iterable - 1 do
-            coroutine.yield(__TS__StringAccess(iterable, index))
-        end
-    elseif iterable.____coroutine ~= nil then
-        local co = iterable.____coroutine
-        while true do
-            local status, value = coroutine.resume(co)
-            if not status then
-                error(value, 0)
-            end
-            if coroutine.status(co) == "dead" then
-                return value
-            else
-                coroutine.yield(value)
-            end
-        end
-    elseif iterable[Symbol.iterator] then
-        local iterator = iterable[Symbol.iterator](iterable)
-        while true do
-            local result = iterator:next()
-            if result.done then
-                return result.value
-            else
-                coroutine.yield(result.value)
-            end
-        end
-    else
-        for ____, value in ipairs(iterable) do
-            coroutine.yield(value)
-        end
-    end
-end
-
-local __TS__Generator
-do
-    local function generatorIterator(self)
-        return self
-    end
-    local function generatorNext(self, ...)
-        local co = self.____coroutine
-        if coroutine.status(co) == "dead" then
-            return {done = true}
-        end
-        local status, value = coroutine.resume(co, ...)
-        if not status then
-            error(value, 0)
-        end
-        return {
-            value = value,
-            done = coroutine.status(co) == "dead"
-        }
-    end
-    function __TS__Generator(fn)
-        return function(...)
-            local args = {...}
-            local argsLength = __TS__CountVarargs(...)
-            return {
-                ____coroutine = coroutine.create(function() return fn(__TS__Unpack(args, 1, argsLength)) end),
-                [Symbol.iterator] = generatorIterator,
-                next = generatorNext
-            }
-        end
-    end
-end
-
-local function __TS__InstanceOfObject(value)
-    local valueType = type(value)
-    return valueType == "table" or valueType == "function"
-end
-
-local function __TS__LuaIteratorSpread(self, state, firstKey)
-    local results = {}
-    local key, value = self(state, firstKey)
-    while key do
-        results[#results + 1] = {key, value}
-        key, value = self(state, key)
-    end
-    return __TS__Unpack(results)
-end
-
-local Map
-do
-    Map = __TS__Class()
-    Map.name = "Map"
-    function Map.prototype.____constructor(self, entries)
-        self[Symbol.toStringTag] = "Map"
-        self.items = {}
-        self.size = 0
-        self.nextKey = {}
-        self.previousKey = {}
-        if entries == nil then
-            return
-        end
-        local iterable = entries
-        if iterable[Symbol.iterator] then
-            local iterator = iterable[Symbol.iterator](iterable)
-            while true do
-                local result = iterator:next()
-                if result.done then
-                    break
-                end
-                local value = result.value
-                self:set(value[1], value[2])
-            end
-        else
-            local array = entries
-            for ____, kvp in ipairs(array) do
-                self:set(kvp[1], kvp[2])
-            end
-        end
-    end
-    function Map.prototype.clear(self)
-        self.items = {}
-        self.nextKey = {}
-        self.previousKey = {}
-        self.firstKey = nil
-        self.lastKey = nil
-        self.size = 0
-    end
-    function Map.prototype.delete(self, key)
-        local contains = self:has(key)
-        if contains then
-            self.size = self.size - 1
-            local next = self.nextKey[key]
-            local previous = self.previousKey[key]
-            if next and previous then
-                self.nextKey[previous] = next
-                self.previousKey[next] = previous
-            elseif next then
-                self.firstKey = next
-                self.previousKey[next] = nil
-            elseif previous then
-                self.lastKey = previous
-                self.nextKey[previous] = nil
-            else
-                self.firstKey = nil
-                self.lastKey = nil
-            end
-            self.nextKey[key] = nil
-            self.previousKey[key] = nil
-        end
-        self.items[key] = nil
-        return contains
-    end
-    function Map.prototype.forEach(self, callback)
-        for ____, key in __TS__Iterator(self:keys()) do
-            callback(nil, self.items[key], key, self)
-        end
-    end
-    function Map.prototype.get(self, key)
-        return self.items[key]
-    end
-    function Map.prototype.has(self, key)
-        return self.nextKey[key] ~= nil or self.lastKey == key
-    end
-    function Map.prototype.set(self, key, value)
-        local isNewValue = not self:has(key)
-        if isNewValue then
-            self.size = self.size + 1
-        end
-        self.items[key] = value
-        if self.firstKey == nil then
-            self.firstKey = key
-            self.lastKey = key
-        elseif isNewValue then
-            self.nextKey[self.lastKey] = key
-            self.previousKey[key] = self.lastKey
-            self.lastKey = key
-        end
-        return self
-    end
-    Map.prototype[Symbol.iterator] = function(self)
-        return self:entries()
-    end
-    function Map.prototype.entries(self)
-        local items = self.items
-        local nextKey = self.nextKey
-        local key = self.firstKey
-        return {
-            [Symbol.iterator] = function(self)
-                return self
-            end,
-            next = function(self)
-                local result = {done = not key, value = {key, items[key]}}
-                key = nextKey[key]
-                return result
-            end
-        }
-    end
-    function Map.prototype.keys(self)
-        local nextKey = self.nextKey
-        local key = self.firstKey
-        return {
-            [Symbol.iterator] = function(self)
-                return self
-            end,
-            next = function(self)
-                local result = {done = not key, value = key}
-                key = nextKey[key]
-                return result
-            end
-        }
-    end
-    function Map.prototype.values(self)
-        local items = self.items
-        local nextKey = self.nextKey
-        local key = self.firstKey
-        return {
-            [Symbol.iterator] = function(self)
-                return self
-            end,
-            next = function(self)
-                local result = {done = not key, value = items[key]}
-                key = nextKey[key]
-                return result
-            end
-        }
-    end
-    Map[Symbol.species] = Map
-end
-
-local __TS__Match = string.match
-
-local __TS__MathAtan2 = math.atan2 or math.atan
-
-local __TS__MathModf = math.modf
-
-local function __TS__MathSign(val)
-    if val > 0 then
-        return 1
-    elseif val < 0 then
-        return -1
-    end
-    return 0
-end
-
-local function __TS__Modulo50(a, b)
-    return a - math.floor(a / b) * b
-end
-
-local function __TS__Number(value)
-    local valueType = type(value)
-    if valueType == "number" then
-        return value
-    elseif valueType == "string" then
-        local numberValue = tonumber(value)
-        if numberValue then
-            return numberValue
-        end
-        if value == "Infinity" then
-            return math.huge
-        end
-        if value == "-Infinity" then
-            return -math.huge
-        end
-        local stringWithoutSpaces = string.gsub(value, "%s", "")
-        if stringWithoutSpaces == "" then
-            return 0
-        end
-        return 0 / 0
-    elseif valueType == "boolean" then
-        return value and 1 or 0
-    else
-        return 0 / 0
-    end
-end
-
-local function __TS__NumberIsFinite(value)
-    return type(value) == "number" and value == value and value ~= math.huge and value ~= -math.huge
-end
-
-local function __TS__NumberIsNaN(value)
-    return value ~= value
-end
-
-local __TS__NumberToString
-do
-    local radixChars = "0123456789abcdefghijklmnopqrstuvwxyz"
-    function __TS__NumberToString(self, radix)
-        if radix == nil or radix == 10 or self == math.huge or self == -math.huge or self ~= self then
-            return tostring(self)
-        end
-        radix = math.floor(radix)
-        if radix < 2 or radix > 36 then
-            error("toString() radix argument must be between 2 and 36", 0)
-        end
-        local integer, fraction = __TS__MathModf(math.abs(self))
-        local result = ""
-        if radix == 8 then
-            result = string.format("%o", integer)
-        elseif radix == 16 then
-            result = string.format("%x", integer)
-        else
-            repeat
-                do
-                    result = __TS__StringAccess(radixChars, integer % radix) .. result
-                    integer = math.floor(integer / radix)
-                end
-            until not (integer ~= 0)
-        end
-        if fraction ~= 0 then
-            result = result .. "."
-            local delta = 1e-16
-            repeat
-                do
-                    fraction = fraction * radix
-                    delta = delta * radix
-                    local digit = math.floor(fraction)
-                    result = result .. __TS__StringAccess(radixChars, digit)
-                    fraction = fraction - digit
-                end
-            until not (fraction >= delta)
-        end
-        if self < 0 then
-            result = "-" .. result
-        end
-        return result
-    end
-end
-
-local function __TS__ObjectDefineProperty(target, key, desc)
-    local ____temp_0
-    if type(key) == "number" then
-        ____temp_0 = key + 1
-    else
-        ____temp_0 = key
-    end
-    local luaKey = ____temp_0
-    local value = rawget(target, luaKey)
-    local hasGetterOrSetter = desc.get ~= nil or desc.set ~= nil
-    local descriptor
-    if hasGetterOrSetter then
-        if value ~= nil then
-            error(
-                "Cannot redefine property: " .. tostring(key),
-                0
-            )
-        end
-        descriptor = desc
-    else
-        local valueExists = value ~= nil
-        local ____desc_set_5 = desc.set
-        local ____desc_get_6 = desc.get
-        local ____temp_1
-        if desc.configurable ~= nil then
-            ____temp_1 = desc.configurable
-        else
-            ____temp_1 = valueExists
-        end
-        local ____temp_2
-        if desc.enumerable ~= nil then
-            ____temp_2 = desc.enumerable
-        else
-            ____temp_2 = valueExists
-        end
-        local ____temp_3
-        if desc.writable ~= nil then
-            ____temp_3 = desc.writable
-        else
-            ____temp_3 = valueExists
-        end
-        local ____temp_4
-        if desc.value ~= nil then
-            ____temp_4 = desc.value
-        else
-            ____temp_4 = value
-        end
-        descriptor = {
-            set = ____desc_set_5,
-            get = ____desc_get_6,
-            configurable = ____temp_1,
-            enumerable = ____temp_2,
-            writable = ____temp_3,
-            value = ____temp_4
-        }
-    end
-    __TS__SetDescriptor(target, luaKey, descriptor)
-    return target
-end
-
-local function __TS__ObjectEntries(obj)
-    local result = {}
-    local len = 0
-    for key in pairs(obj) do
-        len = len + 1
-        result[len] = {key, obj[key]}
-    end
-    return result
-end
-
-local function __TS__ObjectFromEntries(entries)
-    local obj = {}
-    local iterable = entries
-    if iterable[Symbol.iterator] then
-        local iterator = iterable[Symbol.iterator](iterable)
-        while true do
-            local result = iterator:next()
-            if result.done then
-                break
-            end
-            local value = result.value
-            obj[value[1]] = value[2]
-        end
-    else
-        for ____, entry in ipairs(entries) do
-            obj[entry[1]] = entry[2]
-        end
-    end
-    return obj
-end
-
-local function __TS__ObjectKeys(obj)
-    local result = {}
-    local len = 0
-    for key in pairs(obj) do
-        len = len + 1
-        result[len] = key
-    end
-    return result
-end
-
-local function __TS__ObjectRest(target, usedProperties)
-    local result = {}
-    for property in pairs(target) do
-        if not usedProperties[property] then
-            result[property] = target[property]
-        end
-    end
-    return result
-end
-
-local function __TS__ObjectValues(obj)
-    local result = {}
-    local len = 0
-    for key in pairs(obj) do
-        len = len + 1
-        result[len] = obj[key]
-    end
-    return result
-end
-
-local function __TS__ParseFloat(numberString)
-    local infinityMatch = __TS__Match(numberString, "^%s*(-?Infinity)")
-    if infinityMatch then
-        local ____temp_0
-        if __TS__StringAccess(infinityMatch, 0) == "-" then
-            ____temp_0 = -math.huge
-        else
-            ____temp_0 = math.huge
-        end
-        return ____temp_0
-    end
-    local number = tonumber(__TS__Match(numberString, "^%s*(-?%d+%.?%d*)"))
-    local ____number_1 = number
-    if ____number_1 == nil then
-        ____number_1 = 0 / 0
-    end
-    return ____number_1
-end
-
-local function __TS__StringSubstr(self, from, length)
-    if from ~= from then
-        from = 0
-    end
-    if length ~= nil then
-        if length ~= length or length <= 0 then
-            return ""
-        end
-        length = length + from
-    end
-    if from >= 0 then
-        from = from + 1
-    end
-    return string.sub(self, from, length)
-end
-
-local function __TS__StringSubstring(self, start, ____end)
-    if ____end ~= ____end then
-        ____end = 0
-    end
-    if ____end ~= nil and start > ____end then
-        start, ____end = ____end, start
-    end
-    if start >= 0 then
-        start = start + 1
-    else
-        start = 1
-    end
-    if ____end ~= nil and ____end < 0 then
-        ____end = 0
-    end
-    return string.sub(self, start, ____end)
-end
-
-local __TS__ParseInt
-do
-    local parseIntBasePattern = "0123456789aAbBcCdDeEfFgGhHiIjJkKlLmMnNoOpPqQrRsStTvVwWxXyYzZ"
-    function __TS__ParseInt(numberString, base)
-        if base == nil then
-            base = 10
-            local hexMatch = __TS__Match(numberString, "^%s*-?0[xX]")
-            if hexMatch then
-                base = 16
-                local ____TS__Match_result__0_0
-                if __TS__Match(hexMatch, "-") then
-                    ____TS__Match_result__0_0 = "-" .. __TS__StringSubstr(numberString, #hexMatch)
-                else
-                    ____TS__Match_result__0_0 = __TS__StringSubstr(numberString, #hexMatch)
-                end
-                numberString = ____TS__Match_result__0_0
-            end
-        end
-        if base < 2 or base > 36 then
-            return 0 / 0
-        end
-        local ____temp_1
-        if base <= 10 then
-            ____temp_1 = __TS__StringSubstring(parseIntBasePattern, 0, base)
-        else
-            ____temp_1 = __TS__StringSubstr(parseIntBasePattern, 0, 10 + 2 * (base - 10))
-        end
-        local allowedDigits = ____temp_1
-        local pattern = ("^%s*(-?[" .. allowedDigits) .. "]*)"
-        local number = tonumber(
-            __TS__Match(numberString, pattern),
-            base
-        )
-        if number == nil then
-            return 0 / 0
-        end
-        if number >= 0 then
-            return math.floor(number)
-        else
-            return math.ceil(number)
-        end
-    end
-end
-
-local function __TS__PromiseAll(iterable)
-    local results = {}
-    local toResolve = {}
-    local numToResolve = 0
-    local i = 0
-    for ____, item in __TS__Iterator(iterable) do
-        if __TS__InstanceOf(item, __TS__Promise) then
-            if item.state == 1 then
-                results[i + 1] = item.value
-            elseif item.state == 2 then
-                return __TS__Promise.reject(item.rejectionReason)
-            else
-                numToResolve = numToResolve + 1
-                toResolve[i] = item
-            end
-        else
-            results[i + 1] = item
-        end
-        i = i + 1
-    end
-    if numToResolve == 0 then
-        return __TS__Promise.resolve(results)
-    end
-    return __TS__New(
-        __TS__Promise,
-        function(____, resolve, reject)
-            for index, promise in pairs(toResolve) do
-                promise["then"](
-                    promise,
-                    function(____, data)
-                        results[index + 1] = data
-                        numToResolve = numToResolve - 1
-                        if numToResolve == 0 then
-                            resolve(nil, results)
-                        end
-                    end,
-                    function(____, reason)
-                        reject(nil, reason)
-                    end
-                )
-            end
-        end
-    )
-end
-
-local function __TS__PromiseAllSettled(iterable)
-    local results = {}
-    local toResolve = {}
-    local numToResolve = 0
-    local i = 0
-    for ____, item in __TS__Iterator(iterable) do
-        if __TS__InstanceOf(item, __TS__Promise) then
-            if item.state == 1 then
-                results[i + 1] = {status = "fulfilled", value = item.value}
-            elseif item.state == 2 then
-                results[i + 1] = {status = "rejected", reason = item.rejectionReason}
-            else
-                numToResolve = numToResolve + 1
-                toResolve[i] = item
-            end
-        else
-            results[i + 1] = {status = "fulfilled", value = item}
-        end
-        i = i + 1
-    end
-    if numToResolve == 0 then
-        return __TS__Promise.resolve(results)
-    end
-    return __TS__New(
-        __TS__Promise,
-        function(____, resolve)
-            for index, promise in pairs(toResolve) do
-                promise["then"](
-                    promise,
-                    function(____, data)
-                        results[index + 1] = {status = "fulfilled", value = data}
-                        numToResolve = numToResolve - 1
-                        if numToResolve == 0 then
-                            resolve(nil, results)
-                        end
-                    end,
-                    function(____, reason)
-                        results[index + 1] = {status = "rejected", reason = reason}
-                        numToResolve = numToResolve - 1
-                        if numToResolve == 0 then
-                            resolve(nil, results)
-                        end
-                    end
-                )
-            end
-        end
-    )
-end
-
-local function __TS__PromiseAny(iterable)
-    local rejections = {}
-    local pending = {}
-    for ____, item in __TS__Iterator(iterable) do
-        if __TS__InstanceOf(item, __TS__Promise) then
-            if item.state == 1 then
-                return __TS__Promise.resolve(item.value)
-            elseif item.state == 2 then
-                rejections[#rejections + 1] = item.rejectionReason
-            else
-                pending[#pending + 1] = item
-            end
-        else
-            return __TS__Promise.resolve(item)
-        end
-    end
-    if #pending == 0 then
-        return __TS__Promise.reject("No promises to resolve with .any()")
-    end
-    local numResolved = 0
-    return __TS__New(
-        __TS__Promise,
-        function(____, resolve, reject)
-            for ____, promise in ipairs(pending) do
-                promise["then"](
-                    promise,
-                    function(____, data)
-                        resolve(nil, data)
-                    end,
-                    function(____, reason)
-                        rejections[#rejections + 1] = reason
-                        numResolved = numResolved + 1
-                        if numResolved == #pending then
-                            reject(nil, {name = "AggregateError", message = "All Promises rejected", errors = rejections})
-                        end
-                    end
-                )
-            end
-        end
-    )
-end
-
-local function __TS__PromiseRace(iterable)
-    local pending = {}
-    for ____, item in __TS__Iterator(iterable) do
-        if __TS__InstanceOf(item, __TS__Promise) then
-            if item.state == 1 then
-                return __TS__Promise.resolve(item.value)
-            elseif item.state == 2 then
-                return __TS__Promise.reject(item.rejectionReason)
-            else
-                pending[#pending + 1] = item
-            end
-        else
-            return __TS__Promise.resolve(item)
-        end
-    end
-    return __TS__New(
-        __TS__Promise,
-        function(____, resolve, reject)
-            for ____, promise in ipairs(pending) do
-                promise["then"](
-                    promise,
-                    function(____, value) return resolve(nil, value) end,
-                    function(____, reason) return reject(nil, reason) end
-                )
-            end
-        end
-    )
-end
-
-local Set
-do
-    Set = __TS__Class()
-    Set.name = "Set"
-    function Set.prototype.____constructor(self, values)
-        self[Symbol.toStringTag] = "Set"
-        self.size = 0
-        self.nextKey = {}
-        self.previousKey = {}
-        if values == nil then
-            return
-        end
-        local iterable = values
-        if iterable[Symbol.iterator] then
-            local iterator = iterable[Symbol.iterator](iterable)
-            while true do
-                local result = iterator:next()
-                if result.done then
-                    break
-                end
-                self:add(result.value)
-            end
-        else
-            local array = values
-            for ____, value in ipairs(array) do
-                self:add(value)
-            end
-        end
-    end
-    function Set.prototype.add(self, value)
-        local isNewValue = not self:has(value)
-        if isNewValue then
-            self.size = self.size + 1
-        end
-        if self.firstKey == nil then
-            self.firstKey = value
-            self.lastKey = value
-        elseif isNewValue then
-            self.nextKey[self.lastKey] = value
-            self.previousKey[value] = self.lastKey
-            self.lastKey = value
-        end
-        return self
-    end
-    function Set.prototype.clear(self)
-        self.nextKey = {}
-        self.previousKey = {}
-        self.firstKey = nil
-        self.lastKey = nil
-        self.size = 0
-    end
-    function Set.prototype.delete(self, value)
-        local contains = self:has(value)
-        if contains then
-            self.size = self.size - 1
-            local next = self.nextKey[value]
-            local previous = self.previousKey[value]
-            if next and previous then
-                self.nextKey[previous] = next
-                self.previousKey[next] = previous
-            elseif next then
-                self.firstKey = next
-                self.previousKey[next] = nil
-            elseif previous then
-                self.lastKey = previous
-                self.nextKey[previous] = nil
-            else
-                self.firstKey = nil
-                self.lastKey = nil
-            end
-            self.nextKey[value] = nil
-            self.previousKey[value] = nil
-        end
-        return contains
-    end
-    function Set.prototype.forEach(self, callback)
-        for ____, key in __TS__Iterator(self:keys()) do
-            callback(nil, key, key, self)
-        end
-    end
-    function Set.prototype.has(self, value)
-        return self.nextKey[value] ~= nil or self.lastKey == value
-    end
-    Set.prototype[Symbol.iterator] = function(self)
-        return self:values()
-    end
-    function Set.prototype.entries(self)
-        local nextKey = self.nextKey
-        local key = self.firstKey
-        return {
-            [Symbol.iterator] = function(self)
-                return self
-            end,
-            next = function(self)
-                local result = {done = not key, value = {key, key}}
-                key = nextKey[key]
-                return result
-            end
-        }
-    end
-    function Set.prototype.keys(self)
-        local nextKey = self.nextKey
-        local key = self.firstKey
-        return {
-            [Symbol.iterator] = function(self)
-                return self
-            end,
-            next = function(self)
-                local result = {done = not key, value = key}
-                key = nextKey[key]
-                return result
-            end
-        }
-    end
-    function Set.prototype.values(self)
-        local nextKey = self.nextKey
-        local key = self.firstKey
-        return {
-            [Symbol.iterator] = function(self)
-                return self
-            end,
-            next = function(self)
-                local result = {done = not key, value = key}
-                key = nextKey[key]
-                return result
-            end
-        }
-    end
-    Set[Symbol.species] = Set
-end
-
-local function __TS__SparseArrayNew(...)
-    local sparseArray = {...}
-    sparseArray.sparseLength = __TS__CountVarargs(...)
-    return sparseArray
-end
-
-local function __TS__SparseArrayPush(sparseArray, ...)
-    local args = {...}
-    local argsLen = __TS__CountVarargs(...)
-    local listLen = sparseArray.sparseLength
-    for i = 1, argsLen do
-        sparseArray[listLen + i] = args[i]
-    end
-    sparseArray.sparseLength = listLen + argsLen
-end
-
-local function __TS__SparseArraySpread(sparseArray)
-    local ____unpack_0 = unpack
-    if ____unpack_0 == nil then
-        ____unpack_0 = table.unpack
-    end
-    local _unpack = ____unpack_0
-    return _unpack(sparseArray, 1, sparseArray.sparseLength)
-end
-
-local WeakMap
-do
-    WeakMap = __TS__Class()
-    WeakMap.name = "WeakMap"
-    function WeakMap.prototype.____constructor(self, entries)
-        self[Symbol.toStringTag] = "WeakMap"
-        self.items = {}
-        setmetatable(self.items, {__mode = "k"})
-        if entries == nil then
-            return
-        end
-        local iterable = entries
-        if iterable[Symbol.iterator] then
-            local iterator = iterable[Symbol.iterator](iterable)
-            while true do
-                local result = iterator:next()
-                if result.done then
-                    break
-                end
-                local value = result.value
-                self.items[value[1]] = value[2]
-            end
-        else
-            for ____, kvp in ipairs(entries) do
-                self.items[kvp[1]] = kvp[2]
-            end
-        end
-    end
-    function WeakMap.prototype.delete(self, key)
-        local contains = self:has(key)
-        self.items[key] = nil
-        return contains
-    end
-    function WeakMap.prototype.get(self, key)
-        return self.items[key]
-    end
-    function WeakMap.prototype.has(self, key)
-        return self.items[key] ~= nil
-    end
-    function WeakMap.prototype.set(self, key, value)
-        self.items[key] = value
-        return self
-    end
-    WeakMap[Symbol.species] = WeakMap
-end
-
-local WeakSet
-do
-    WeakSet = __TS__Class()
-    WeakSet.name = "WeakSet"
-    function WeakSet.prototype.____constructor(self, values)
-        self[Symbol.toStringTag] = "WeakSet"
-        self.items = {}
-        setmetatable(self.items, {__mode = "k"})
-        if values == nil then
-            return
-        end
-        local iterable = values
-        if iterable[Symbol.iterator] then
-            local iterator = iterable[Symbol.iterator](iterable)
-            while true do
-                local result = iterator:next()
-                if result.done then
-                    break
-                end
-                self.items[result.value] = true
-            end
-        else
-            for ____, value in ipairs(values) do
-                self.items[value] = true
-            end
-        end
-    end
-    function WeakSet.prototype.add(self, value)
-        self.items[value] = true
-        return self
-    end
-    function WeakSet.prototype.delete(self, value)
-        local contains = self:has(value)
-        self.items[value] = nil
-        return contains
-    end
-    function WeakSet.prototype.has(self, value)
-        return self.items[value] == true
-    end
-    WeakSet[Symbol.species] = WeakSet
-end
-
-local function __TS__SourceMapTraceBack(fileName, sourceMap)
-    _G.__TS__sourcemap = _G.__TS__sourcemap or ({})
-    _G.__TS__sourcemap[fileName] = sourceMap
-    if _G.__TS__originalTraceback == nil then
-        local originalTraceback = debug.traceback
-        _G.__TS__originalTraceback = originalTraceback
-        debug.traceback = function(thread, message, level)
-            local trace
-            if thread == nil and message == nil and level == nil then
-                trace = originalTraceback()
-            elseif __TS__StringIncludes(_VERSION, "Lua 5.0") then
-                trace = originalTraceback((("[Level " .. tostring(level)) .. "] ") .. message)
-            else
-                trace = originalTraceback(thread, message, level)
-            end
-            if type(trace) ~= "string" then
-                return trace
-            end
-            local function replacer(____, file, srcFile, line)
-                local fileSourceMap = _G.__TS__sourcemap[file]
-                if fileSourceMap and fileSourceMap[line] then
-                    local data = fileSourceMap[line]
-                    if type(data) == "number" then
-                        return (srcFile .. ":") .. tostring(data)
-                    end
-                    return (tostring(data.file) .. ":") .. tostring(data.line)
-                end
-                return (file .. ":") .. line
-            end
-            local result = string.gsub(
-                trace,
-                "(%S+)%.lua:(%d+)",
-                function(file, line) return replacer(nil, file .. ".lua", file .. ".ts", line) end
-            )
-            local function stringReplacer(____, file, line)
-                local fileSourceMap = _G.__TS__sourcemap[file]
-                if fileSourceMap and fileSourceMap[line] then
-                    local chunkName = __TS__Match(file, "%[string \"([^\"]+)\"%]")
-                    local sourceName = string.gsub(chunkName, ".lua$", ".ts")
-                    local data = fileSourceMap[line]
-                    if type(data) == "number" then
-                        return (sourceName .. ":") .. tostring(data)
-                    end
-                    return (tostring(data.file) .. ":") .. tostring(data.line)
-                end
-                return (file .. ":") .. line
-            end
-            result = string.gsub(
-                result,
-                "(%[string \"[^\"]+\"%]):(%d+)",
-                function(file, line) return stringReplacer(nil, file, line) end
-            )
-            return result
-        end
-    end
-end
-
-local function __TS__Spread(iterable)
-    local arr = {}
-    if type(iterable) == "string" then
-        for i = 0, #iterable - 1 do
-            arr[i + 1] = __TS__StringAccess(iterable, i)
-        end
-    else
-        local len = 0
-        for ____, item in __TS__Iterator(iterable) do
-            len = len + 1
-            arr[len] = item
-        end
-    end
-    return __TS__Unpack(arr)
-end
-
-local function __TS__StringCharAt(self, pos)
-    if pos ~= pos then
-        pos = 0
-    end
-    if pos < 0 then
-        return ""
-    end
-    return string.sub(self, pos + 1, pos + 1)
-end
-
-local function __TS__StringCharCodeAt(self, index)
-    if index ~= index then
-        index = 0
-    end
-    if index < 0 then
-        return 0 / 0
-    end
-    local ____string_byte_result_0 = string.byte(self, index + 1)
-    if ____string_byte_result_0 == nil then
-        ____string_byte_result_0 = 0 / 0
-    end
-    return ____string_byte_result_0
-end
-
-local function __TS__StringEndsWith(self, searchString, endPosition)
-    if endPosition == nil or endPosition > #self then
-        endPosition = #self
-    end
-    return string.sub(self, endPosition - #searchString + 1, endPosition) == searchString
-end
-
-local function __TS__StringPadEnd(self, maxLength, fillString)
-    if fillString == nil then
-        fillString = " "
-    end
-    if maxLength ~= maxLength then
-        maxLength = 0
-    end
-    if maxLength == -math.huge or maxLength == math.huge then
-        error("Invalid string length", 0)
-    end
-    if #self >= maxLength or #fillString == 0 then
-        return self
-    end
-    maxLength = maxLength - #self
-    if maxLength > #fillString then
-        fillString = fillString .. string.rep(
-            fillString,
-            math.floor(maxLength / #fillString)
-        )
-    end
-    return self .. string.sub(
-        fillString,
-        1,
-        math.floor(maxLength)
-    )
-end
-
-local function __TS__StringPadStart(self, maxLength, fillString)
-    if fillString == nil then
-        fillString = " "
-    end
-    if maxLength ~= maxLength then
-        maxLength = 0
-    end
-    if maxLength == -math.huge or maxLength == math.huge then
-        error("Invalid string length", 0)
-    end
-    if #self >= maxLength or #fillString == 0 then
-        return self
-    end
-    maxLength = maxLength - #self
-    if maxLength > #fillString then
-        fillString = fillString .. string.rep(
-            fillString,
-            math.floor(maxLength / #fillString)
-        )
-    end
-    return string.sub(
-        fillString,
-        1,
-        math.floor(maxLength)
-    ) .. self
-end
-
-local __TS__StringReplace
-do
-    local sub = string.sub
-    function __TS__StringReplace(source, searchValue, replaceValue)
-        local startPos, endPos = string.find(source, searchValue, nil, true)
-        if not startPos then
-            return source
-        end
-        local before = sub(source, 1, startPos - 1)
-        local ____temp_0
-        if type(replaceValue) == "string" then
-            ____temp_0 = replaceValue
-        else
-            ____temp_0 = replaceValue(nil, searchValue, startPos - 1, source)
-        end
-        local replacement = ____temp_0
-        local after = sub(source, endPos + 1)
-        return (before .. replacement) .. after
-    end
-end
-
-local __TS__StringSplit
-do
-    local sub = string.sub
-    local find = string.find
-    function __TS__StringSplit(source, separator, limit)
-        if limit == nil then
-            limit = 4294967295
-        end
-        if limit == 0 then
-            return {}
-        end
-        local result = {}
-        local resultIndex = 1
-        if separator == nil or separator == "" then
-            for i = 1, #source do
-                result[resultIndex] = sub(source, i, i)
-                resultIndex = resultIndex + 1
-            end
-        else
-            local currentPos = 1
-            while resultIndex <= limit do
-                local startPos, endPos = find(source, separator, currentPos, true)
-                if not startPos then
-                    break
-                end
-                result[resultIndex] = sub(source, currentPos, startPos - 1)
-                resultIndex = resultIndex + 1
-                currentPos = endPos + 1
-            end
-            if resultIndex <= limit then
-                result[resultIndex] = sub(source, currentPos)
-            end
-        end
-        return result
-    end
-end
-
-local __TS__StringReplaceAll
-do
-    local sub = string.sub
-    local find = string.find
-    function __TS__StringReplaceAll(source, searchValue, replaceValue)
-        if type(replaceValue) == "string" then
-            local concat = table.concat(
-                __TS__StringSplit(source, searchValue),
-                replaceValue
-            )
-            if #searchValue == 0 then
-                return (replaceValue .. concat) .. replaceValue
-            end
-            return concat
-        end
-        local parts = {}
-        local partsIndex = 1
-        if #searchValue == 0 then
-            parts[1] = replaceValue(nil, "", 0, source)
-            partsIndex = 2
-            for i = 1, #source do
-                parts[partsIndex] = sub(source, i, i)
-                parts[partsIndex + 1] = replaceValue(nil, "", i, source)
-                partsIndex = partsIndex + 2
-            end
-        else
-            local currentPos = 1
-            while true do
-                local startPos, endPos = find(source, searchValue, currentPos, true)
-                if not startPos then
-                    break
-                end
-                parts[partsIndex] = sub(source, currentPos, startPos - 1)
-                parts[partsIndex + 1] = replaceValue(nil, searchValue, startPos - 1, source)
-                partsIndex = partsIndex + 2
-                currentPos = endPos + 1
-            end
-            parts[partsIndex] = sub(source, currentPos)
-        end
-        return table.concat(parts)
-    end
-end
-
-local function __TS__StringSlice(self, start, ____end)
-    if start == nil or start ~= start then
-        start = 0
-    end
-    if ____end ~= ____end then
-        ____end = 0
-    end
-    if start >= 0 then
-        start = start + 1
-    end
-    if ____end ~= nil and ____end < 0 then
-        ____end = ____end - 1
-    end
-    return string.sub(self, start, ____end)
-end
-
-local function __TS__StringStartsWith(self, searchString, position)
-    if position == nil or position < 0 then
-        position = 0
-    end
-    return string.sub(self, position + 1, #searchString + position) == searchString
-end
-
-local function __TS__StringTrim(self)
-    local result = string.gsub(self, "^[%s ﻿]*(.-)[%s ﻿]*$", "%1")
-    return result
-end
-
-local function __TS__StringTrimEnd(self)
-    local result = string.gsub(self, "[%s ﻿]*$", "")
-    return result
-end
-
-local function __TS__StringTrimStart(self)
-    local result = string.gsub(self, "^[%s ﻿]*", "")
-    return result
-end
-
-local __TS__SymbolRegistryFor, __TS__SymbolRegistryKeyFor
-do
-    local symbolRegistry = {}
-    function __TS__SymbolRegistryFor(key)
-        if not symbolRegistry[key] then
-            symbolRegistry[key] = __TS__Symbol(key)
-        end
-        return symbolRegistry[key]
-    end
-    function __TS__SymbolRegistryKeyFor(sym)
-        for key in pairs(symbolRegistry) do
-            if symbolRegistry[key] == sym then
-                return key
-            end
-        end
-    end
-end
-
-local function __TS__TypeOf(value)
-    local luaType = type(value)
-    if luaType == "table" then
-        return "object"
-    elseif luaType == "nil" then
-        return "undefined"
-    else
-        return luaType
-    end
-end
-
-return {
-  __TS__ArrayConcat = __TS__ArrayConcat,
-  __TS__ArrayEntries = __TS__ArrayEntries,
-  __TS__ArrayEvery = __TS__ArrayEvery,
-  __TS__ArrayFilter = __TS__ArrayFilter,
-  __TS__ArrayForEach = __TS__ArrayForEach,
-  __TS__ArrayFind = __TS__ArrayFind,
-  __TS__ArrayFindIndex = __TS__ArrayFindIndex,
-  __TS__ArrayFrom = __TS__ArrayFrom,
-  __TS__ArrayIncludes = __TS__ArrayIncludes,
-  __TS__ArrayIndexOf = __TS__ArrayIndexOf,
-  __TS__ArrayIsArray = __TS__ArrayIsArray,
-  __TS__ArrayJoin = __TS__ArrayJoin,
-  __TS__ArrayMap = __TS__ArrayMap,
-  __TS__ArrayPush = __TS__ArrayPush,
-  __TS__ArrayPushArray = __TS__ArrayPushArray,
-  __TS__ArrayReduce = __TS__ArrayReduce,
-  __TS__ArrayReduceRight = __TS__ArrayReduceRight,
-  __TS__ArrayReverse = __TS__ArrayReverse,
-  __TS__ArrayUnshift = __TS__ArrayUnshift,
-  __TS__ArraySort = __TS__ArraySort,
-  __TS__ArraySlice = __TS__ArraySlice,
-  __TS__ArraySome = __TS__ArraySome,
-  __TS__ArraySplice = __TS__ArraySplice,
-  __TS__ArrayToObject = __TS__ArrayToObject,
-  __TS__ArrayFlat = __TS__ArrayFlat,
-  __TS__ArrayFlatMap = __TS__ArrayFlatMap,
-  __TS__ArraySetLength = __TS__ArraySetLength,
-  __TS__AsyncAwaiter = __TS__AsyncAwaiter,
-  __TS__Await = __TS__Await,
-  __TS__Class = __TS__Class,
-  __TS__ClassExtends = __TS__ClassExtends,
-  __TS__CloneDescriptor = __TS__CloneDescriptor,
-  __TS__CountVarargs = __TS__CountVarargs,
-  __TS__Decorate = __TS__Decorate,
-  __TS__DecorateParam = __TS__DecorateParam,
-  __TS__Delete = __TS__Delete,
-  __TS__DelegatedYield = __TS__DelegatedYield,
-  Error = Error,
-  RangeError = RangeError,
-  ReferenceError = ReferenceError,
-  SyntaxError = SyntaxError,
-  TypeError = TypeError,
-  URIError = URIError,
-  __TS__FunctionBind = __TS__FunctionBind,
-  __TS__Generator = __TS__Generator,
-  __TS__InstanceOf = __TS__InstanceOf,
-  __TS__InstanceOfObject = __TS__InstanceOfObject,
-  __TS__Iterator = __TS__Iterator,
-  __TS__LuaIteratorSpread = __TS__LuaIteratorSpread,
-  Map = Map,
-  __TS__Match = __TS__Match,
-  __TS__MathAtan2 = __TS__MathAtan2,
-  __TS__MathModf = __TS__MathModf,
-  __TS__MathSign = __TS__MathSign,
-  __TS__Modulo50 = __TS__Modulo50,
-  __TS__New = __TS__New,
-  __TS__Number = __TS__Number,
-  __TS__NumberIsFinite = __TS__NumberIsFinite,
-  __TS__NumberIsNaN = __TS__NumberIsNaN,
-  __TS__NumberToString = __TS__NumberToString,
-  __TS__ObjectAssign = __TS__ObjectAssign,
-  __TS__ObjectDefineProperty = __TS__ObjectDefineProperty,
-  __TS__ObjectEntries = __TS__ObjectEntries,
-  __TS__ObjectFromEntries = __TS__ObjectFromEntries,
-  __TS__ObjectGetOwnPropertyDescriptor = __TS__ObjectGetOwnPropertyDescriptor,
-  __TS__ObjectGetOwnPropertyDescriptors = __TS__ObjectGetOwnPropertyDescriptors,
-  __TS__ObjectKeys = __TS__ObjectKeys,
-  __TS__ObjectRest = __TS__ObjectRest,
-  __TS__ObjectValues = __TS__ObjectValues,
-  __TS__ParseFloat = __TS__ParseFloat,
-  __TS__ParseInt = __TS__ParseInt,
-  __TS__Promise = __TS__Promise,
-  __TS__PromiseAll = __TS__PromiseAll,
-  __TS__PromiseAllSettled = __TS__PromiseAllSettled,
-  __TS__PromiseAny = __TS__PromiseAny,
-  __TS__PromiseRace = __TS__PromiseRace,
-  Set = Set,
-  __TS__SetDescriptor = __TS__SetDescriptor,
-  __TS__SparseArrayNew = __TS__SparseArrayNew,
-  __TS__SparseArrayPush = __TS__SparseArrayPush,
-  __TS__SparseArraySpread = __TS__SparseArraySpread,
-  WeakMap = WeakMap,
-  WeakSet = WeakSet,
-  __TS__SourceMapTraceBack = __TS__SourceMapTraceBack,
-  __TS__Spread = __TS__Spread,
-  __TS__StringAccess = __TS__StringAccess,
-  __TS__StringCharAt = __TS__StringCharAt,
-  __TS__StringCharCodeAt = __TS__StringCharCodeAt,
-  __TS__StringEndsWith = __TS__StringEndsWith,
-  __TS__StringIncludes = __TS__StringIncludes,
-  __TS__StringPadEnd = __TS__StringPadEnd,
-  __TS__StringPadStart = __TS__StringPadStart,
-  __TS__StringReplace = __TS__StringReplace,
-  __TS__StringReplaceAll = __TS__StringReplaceAll,
-  __TS__StringSlice = __TS__StringSlice,
-  __TS__StringSplit = __TS__StringSplit,
-  __TS__StringStartsWith = __TS__StringStartsWith,
-  __TS__StringSubstr = __TS__StringSubstr,
-  __TS__StringSubstring = __TS__StringSubstring,
-  __TS__StringTrim = __TS__StringTrim,
-  __TS__StringTrimEnd = __TS__StringTrimEnd,
-  __TS__StringTrimStart = __TS__StringTrimStart,
-  __TS__Symbol = __TS__Symbol,
-  Symbol = Symbol,
-  __TS__SymbolRegistryFor = __TS__SymbolRegistryFor,
-  __TS__SymbolRegistryKeyFor = __TS__SymbolRegistryKeyFor,
-  __TS__TypeOf = __TS__TypeOf,
-  __TS__Unpack = __TS__Unpack
-}
- end,
-["config"] = function(...) 
-local ____lualib = require("lualib_bundle")
-local __TS__Class = ____lualib.__TS__Class
-local __TS__ObjectAssign = ____lualib.__TS__ObjectAssign
-local __TS__New = ____lualib.__TS__New
 local ____exports = {}
-local Config = __TS__Class()
-Config.name = "Config"
-function Config.prototype.____constructor(self)
-    self.disable_commands = false
-    self.debug = false
-    self.server = {}
-    self.go_to_source_definition = {fallback = true}
-end
-function Config.prototype.setup(self, userOpts)
-    __TS__ObjectAssign(
-        self,
-        vim.tbl_deep_extend("force", self, userOpts)
-    )
-end
-____exports.config = __TS__New(Config)
-____exports.setupConfig = function(userOpts)
-    ____exports.config:setup(userOpts)
-end
-return ____exports
- end,
-["types.methods"] = function(...) 
---[[ Generated with https://github.com/TypeScriptToLua/TypeScriptToLua ]]
-local ____exports = {}
-____exports.Methods = Methods or ({})
-____exports.Methods.CODE_ACTION = "textDocument/codeAction"
-____exports.Methods.DEFINITION = "textDocument/definition"
-____exports.Methods.EXECUTE_COMMAND = "workspace/executeCommand"
-____exports.TypescriptMethods = TypescriptMethods or ({})
-____exports.TypescriptMethods.RENAME = "_typescript.rename"
-return ____exports
- end,
-["types.workspace-commands"] = function(...) 
---[[ Generated with https://github.com/TypeScriptToLua/TypeScriptToLua ]]
-local ____exports = {}
-____exports.WorkspaceCommands = WorkspaceCommands or ({})
-____exports.WorkspaceCommands.APPLY_RENAME_FILE = "_typescript.applyRenameFile"
-____exports.WorkspaceCommands.GO_TO_SOURCE_DEFINITION = "_typescript.goToSourceDefinition"
-return ____exports
- end,
-["utils"] = function(...) 
---[[ Generated with https://github.com/TypeScriptToLua/TypeScriptToLua ]]
-local ____exports = {}
-local ____config = require("config")
-local config = ____config.config
-____exports.debugLog = function(...)
-    if not config.debug then
-        return
-    end
-    print(...)
-end
-____exports.getClient = function(bufnr)
-    for ____, client in pairs(vim.lsp.get_active_clients({bufnr = bufnr})) do
-        if client.name == "tsserver" then
-            return client
-        end
-    end
-end
-____exports.resolveHandler = function(bufnr, method)
-    local client = ____exports.getClient(bufnr)
-    if not client then
-        return
-    end
-    return client.handlers[method] or vim.lsp.handlers[method]
-end
-return ____exports
- end,
-["execute-command"] = function(...) 
---[[ Generated with https://github.com/TypeScriptToLua/TypeScriptToLua ]]
-local ____exports = {}
-local ____methods = require("types.methods")
-local Methods = ____methods.Methods
-local ____utils = require("utils")
-local getClient = ____utils.getClient
-____exports.executeCommand = function(bufnr, params, callback)
-    local client = getClient(bufnr)
-    if not client then
-        return false
-    end
-    return client.request(Methods.EXECUTE_COMMAND, params, callback)
-end
-return ____exports
- end,
-["go-to-source-definition"] = function(...) 
---[[ Generated with https://github.com/TypeScriptToLua/TypeScriptToLua ]]
-local ____exports = {}
-local ____execute_2Dcommand = require("execute-command")
-local executeCommand = ____execute_2Dcommand.executeCommand
-local ____methods = require("types.methods")
-local Methods = ____methods.Methods
-local ____workspace_2Dcommands = require("types.workspace-commands")
-local WorkspaceCommands = ____workspace_2Dcommands.WorkspaceCommands
-local ____utils = require("utils")
-local getClient = ____utils.getClient
-local resolveHandler = ____utils.resolveHandler
-____exports.goToSourceDefinition = function(winnr, ____bindingPattern0)
-    local fallback
-    fallback = ____bindingPattern0.fallback
-    local bufnr = vim.api.nvim_win_get_buf(winnr)
-    local client = getClient(bufnr)
-    if not client then
-        return false
-    end
-    local positionParams = vim.lsp.util.make_position_params(winnr, client.offset_encoding)
-    local requestOk = executeCommand(
-        bufnr,
-        {command = WorkspaceCommands.GO_TO_SOURCE_DEFINITION, arguments = {positionParams.textDocument.uri, positionParams.position}},
-        function(...)
-            local args = {...}
-            local handler = resolveHandler(bufnr, Methods.DEFINITION)
-            if not handler then
-                print("failed to go to source definition: could not resolve definition handler")
-                return
-            end
-            local res = args[2] or ({})
-            if vim.tbl_isempty(res) then
-                if fallback == true then
-                    return client.request(Methods.DEFINITION, positionParams, handler, bufnr)
-                end
-                print("failed to go to source definition: no source definitions found")
-                return
-            end
-            handler(unpack(args))
-        end
-    )
-    if not requestOk then
-        print("failed to go to source definition: tsserver request failed")
-    end
-    return requestOk
-end
-return ____exports
- end,
-["rename-file"] = function(...) 
-local ____lualib = require("lualib_bundle")
-local Error = ____lualib.Error
-local RangeError = ____lualib.RangeError
-local ReferenceError = ____lualib.ReferenceError
-local SyntaxError = ____lualib.SyntaxError
-local TypeError = ____lualib.TypeError
-local URIError = ____lualib.URIError
-local __TS__New = ____lualib.__TS__New
-local ____exports = {}
-local ____execute_2Dcommand = require("execute-command")
-local executeCommand = ____execute_2Dcommand.executeCommand
-local ____workspace_2Dcommands = require("types.workspace-commands")
-local WorkspaceCommands = ____workspace_2Dcommands.WorkspaceCommands
-local ____utils = require("utils")
-local debugLog = ____utils.debugLog
-local ____lspconfig = require("lspconfig")
-local util = ____lspconfig.util
-____exports.renameFile = function(source, target, opts)
-    if opts == nil then
-        opts = {}
-    end
-    local sourceBufnr = vim.fn.bufadd(source)
-    vim.fn.bufload(sourceBufnr)
-    if util.path.exists(target) and (opts.force == nil or opts.force == false) then
-        local status = vim.fn.confirm("File exists! Overwrite?", "&Yes\n&No")
-        if status ~= 1 then
-            debugLog("user declined to overrwrite file; aborting")
-            return false
-        end
-    end
-    vim.fn.mkdir(
-        vim.fn.fnamemodify(target, ":p:h"),
-        "p"
-    )
-    debugLog((("sending request to rename source " .. source) .. " to target ") .. target)
-    local requestOk = executeCommand(
-        sourceBufnr,
-        {
-            command = WorkspaceCommands.APPLY_RENAME_FILE,
-            arguments = {{
-                sourceUri = vim.uri_from_fname(source),
-                targetUri = vim.uri_from_fname(target)
-            }}
-        }
-    )
-    if not requestOk then
-        print("failed to rename file: tsserver request failed")
-        return false
-    end
-    if vim.api.nvim_buf_get_option(sourceBufnr, "modified") then
-        vim.api.nvim_buf_call(
-            sourceBufnr,
-            function() return vim.cmd("w!") end
-        )
-    end
-    local didRename, renameError = vim.loop.fs_rename(source, target)
-    if not didRename then
-        error(
-            __TS__New(Error, (((("failed to move " .. source) .. " to ") .. target) .. ": ") .. renameError),
-            0
-        )
-    else
-        debugLog((("successfully renamed source " .. source) .. " to target ") .. target)
-    end
-    local targetBufnr = vim.fn.bufadd(target)
-    vim.api.nvim_buf_set_option(targetBufnr, "buflisted", true)
-    for ____, win in ipairs(vim.api.nvim_list_wins()) do
-        if vim.api.nvim_win_get_buf(win) == sourceBufnr then
-            vim.api.nvim_win_set_buf(win, targetBufnr)
-        end
-    end
-    vim.schedule(function() return vim.api.nvim_buf_delete(sourceBufnr, {force = true}) end)
-    return true
-end
-return ____exports
- end,
-["source-actions"] = function(...) 
-local ____lualib = require("lualib_bundle")
-local __TS__ObjectAssign = ____lualib.__TS__ObjectAssign
-local ____exports = {}
-local ____methods = require("types.methods")
-local Methods = ____methods.Methods
-local ____utils = require("utils")
-local debugLog = ____utils.debugLog
-local getClient = ____utils.getClient
-local SourceActions = SourceActions or ({})
-SourceActions.SourceAddMissingImportsTs = "source.addMissingImports.ts"
-SourceActions.SourceFixAllTs = "source.fixAll.ts"
-SourceActions.SourceRemoveUnusedTs = "source.removeUnused.ts"
-SourceActions.SourceOrganizeImportsTs = "source.organizeImports.ts"
-local function makeCommand(sourceAction)
-    return function(opts)
-        if opts == nil then
-            opts = {}
-        end
-        local bufnr = opts.bufnr or vim.api.nvim_get_current_buf()
-        local client = getClient(bufnr)
-        if not client then
-            return
-        end
-        local params = __TS__ObjectAssign(
-            {},
-            vim.lsp.util.make_range_params(),
-            {context = {
-                only = {sourceAction},
-                diagnostics = vim.diagnostic.get(bufnr)
-            }}
-        )
-        local function applyEdits(res)
-            debugLog(
-                "received response:",
-                vim.inspect(res)
-            )
-            local ____res__0_edit_6 = res[1]
-            if ____res__0_edit_6 ~= nil then
-                ____res__0_edit_6 = ____res__0_edit_6.edit
-            end
-            local ____res__0_edit_documentChanges_4 = ____res__0_edit_6
-            if ____res__0_edit_documentChanges_4 ~= nil then
-                ____res__0_edit_documentChanges_4 = ____res__0_edit_documentChanges_4.documentChanges
-            end
-            local ____res__0_edit_documentChanges__0_2 = ____res__0_edit_documentChanges_4
-            if ____res__0_edit_documentChanges__0_2 ~= nil then
-                ____res__0_edit_documentChanges__0_2 = ____res__0_edit_documentChanges__0_2[1]
-            end
-            local ____res__0_edit_documentChanges__0_edits_0 = ____res__0_edit_documentChanges__0_2
-            if ____res__0_edit_documentChanges__0_edits_0 ~= nil then
-                ____res__0_edit_documentChanges__0_edits_0 = ____res__0_edit_documentChanges__0_edits_0.edits
-            end
-            if ____res__0_edit_documentChanges__0_edits_0 == nil then
-                return
-            end
-            vim.lsp.util.apply_text_edits(res[1].edit.documentChanges[1].edits, bufnr, client.offset_encoding)
-        end
-        debugLog(
-            ("sending source action request for action " .. sourceAction) .. " with params:",
-            vim.inspect(params)
-        )
-        if opts.sync == true then
-            local res = client.request_sync(Methods.CODE_ACTION, params, nil, bufnr)
-            applyEdits(res.result)
-        else
-            client.request(
-                Methods.CODE_ACTION,
-                params,
-                function(_, res) return applyEdits(res or ({})) end,
-                bufnr
-            )
-        end
-    end
-end
-____exports.addMissingImports = makeCommand(SourceActions.SourceAddMissingImportsTs)
-____exports.organizeImports = makeCommand(SourceActions.SourceOrganizeImportsTs)
-____exports.fixAll = makeCommand(SourceActions.SourceFixAllTs)
-____exports.removeUnused = makeCommand(SourceActions.SourceRemoveUnusedTs)
-return ____exports
- end,
-["commands"] = function(...) 
---[[ Generated with https://github.com/TypeScriptToLua/TypeScriptToLua ]]
-local ____exports = {}
-local ____config = require("config")
-local config = ____config.config
-local ____go_2Dto_2Dsource_2Ddefinition = require("go-to-source-definition")
-local goToSourceDefinition = ____go_2Dto_2Dsource_2Ddefinition.goToSourceDefinition
-local ____rename_2Dfile = require("rename-file")
-local renameFile = ____rename_2Dfile.renameFile
-local ____source_2Dactions = require("source-actions")
-local addMissingImports = ____source_2Dactions.addMissingImports
-local fixAll = ____source_2Dactions.fixAll
-local organizeImports = ____source_2Dactions.organizeImports
-local removeUnused = ____source_2Dactions.removeUnused
-____exports.setupCommands = function(bufnr)
-    vim.api.nvim_buf_create_user_command(
-        bufnr,
-        "TypescriptRenameFile",
-        function(opts)
-            local source = vim.api.nvim_buf_get_name(bufnr)
-            vim.ui.input(
-                {prompt = "New path: ", default = source},
-                function(input)
-                    if input == "" or input == source or input == nil then
-                        return
-                    end
-                    renameFile(source, input, {force = opts.bang})
-                end
-            )
-        end,
-        {bang = true}
-    )
-    vim.api.nvim_buf_create_user_command(
-        bufnr,
-        "TypescriptGoToSourceDefinition",
-        function() return goToSourceDefinition(
-            vim.api.nvim_get_current_win(),
-            {fallback = config.go_to_source_definition.fallback}
-        ) end,
-        {}
-    )
-    vim.api.nvim_buf_create_user_command(
-        bufnr,
-        "TypescriptAddMissingImports",
-        function(opts) return addMissingImports({sync = opts.bang, bufnr = bufnr}) end,
-        {bang = true}
-    )
-    vim.api.nvim_buf_create_user_command(
-        bufnr,
-        "TypescriptOrganizeImports",
-        function(opts) return organizeImports({sync = opts.bang, bufnr = bufnr}) end,
-        {bang = true}
-    )
-    vim.api.nvim_buf_create_user_command(
-        bufnr,
-        "TypescriptFixAll",
-        function(opts) return fixAll({sync = opts.bang, bufnr = bufnr}) end,
-        {bang = true}
-    )
-    vim.api.nvim_buf_create_user_command(
-        bufnr,
-        "TypescriptRemoveUnused",
-        function(opts) return removeUnused({sync = opts.bang, bufnr = bufnr}) end,
-        {bang = true}
-    )
-end
-return ____exports
- end,
-["handlers"] = function(...) 
---[[ Generated with https://github.com/TypeScriptToLua/TypeScriptToLua ]]
-local ____exports = {}
-local function isRenameResult(res)
-    return type(res) == "table" and res.position ~= nil
-end
-local function renameHandler(_err, res)
-    if not isRenameResult(res) then
-        return
-    end
-    local bufnr = 0
-    local win = 0
-    local ____res_position_0 = res.position
-    local line = ____res_position_0.line
-    local character = ____res_position_0.character
-    local col = vim.str_byteindex(
-        vim.api.nvim_buf_get_lines(bufnr, line, line + 1, true)[1],
-        character,
-        true
-    )
-    vim.api.nvim_win_set_cursor(win, {line + 1, col})
-    vim.lsp.buf.rename()
-    return res
-end
-____exports.renameHandler = renameHandler
-return ____exports
- end,
-["lsp"] = function(...) 
-local ____lualib = require("lualib_bundle")
-local __TS__ObjectAssign = ____lualib.__TS__ObjectAssign
-local ____exports = {}
-local ____commands = require("commands")
-local setupCommands = ____commands.setupCommands
-local ____config = require("config")
-local config = ____config.config
-local ____handlers = require("handlers")
-local renameHandler = ____handlers.renameHandler
-local ____methods = require("types.methods")
-local TypescriptMethods = ____methods.TypescriptMethods
-local ____lspconfig = require("lspconfig")
-local tsserver = ____lspconfig.tsserver
-____exports.setupLsp = function(overrides)
-    local resolvedConfig = __TS__ObjectAssign({}, config, overrides or ({}))
-    local ____resolvedConfig_server_0 = resolvedConfig.server
-    local on_init = ____resolvedConfig_server_0.on_init
-    local on_attach = ____resolvedConfig_server_0.on_attach
-    local handlers = ____resolvedConfig_server_0.handlers
-    resolvedConfig.server.on_init = function(client, initialize_result)
-        local ____on_init_result_1 = on_init
-        if ____on_init_result_1 ~= nil then
-            ____on_init_result_1 = ____on_init_result_1(client, initialize_result)
-        end
-    end
-    resolvedConfig.server.on_attach = function(client, bufnr)
-        if not config.disable_commands then
-            setupCommands(bufnr)
-        end
-        local ____on_attach_result_3 = on_attach
-        if ____on_attach_result_3 ~= nil then
-            ____on_attach_result_3 = ____on_attach_result_3(client, bufnr)
-        end
-    end
-    local ____resolvedConfig_server_9 = resolvedConfig.server
-    local ____temp_7 = handlers or ({})
-    local ____TypescriptMethods_RENAME_8 = TypescriptMethods.RENAME
-    local ____handlers_TypescriptMethods_RENAME_5 = handlers
-    if ____handlers_TypescriptMethods_RENAME_5 ~= nil then
-        ____handlers_TypescriptMethods_RENAME_5 = ____handlers_TypescriptMethods_RENAME_5[TypescriptMethods.RENAME]
-    end
-    ____resolvedConfig_server_9.handlers = __TS__ObjectAssign({}, ____temp_7, {[____TypescriptMethods_RENAME_8] = ____handlers_TypescriptMethods_RENAME_5 or renameHandler})
-    tsserver.setup(resolvedConfig.server)
-end
-return ____exports
- end,
-["index"] = function(...) 
---[[ Generated with https://github.com/TypeScriptToLua/TypeScriptToLua ]]
-local ____exports = {}
-local ____config = require("config")
+local ____config = require("typescript.config")
 local setupConfig = ____config.setupConfig
-local ____lsp = require("lsp")
+local ____lsp = require("typescript.lsp")
 local setupLsp = ____lsp.setupLsp
 ____exports.setup = function(userOptions)
     setupConfig(userOptions)
     setupLsp()
 end
 do
-    local ____go_2Dto_2Dsource_2Ddefinition = require("go-to-source-definition")
+    local ____go_2Dto_2Dsource_2Ddefinition = require("typescript.go-to-source-definition")
     local goToSourceDefinition = ____go_2Dto_2Dsource_2Ddefinition.goToSourceDefinition
     ____exports.goToSourceDefinition = goToSourceDefinition
 end
 do
-    local ____rename_2Dfile = require("rename-file")
+    local ____rename_2Dfile = require("typescript.rename-file")
     local renameFile = ____rename_2Dfile.renameFile
     ____exports.renameFile = renameFile
 end
-____exports.actions = require("source-actions")
+____exports.actions = require("typescript.source-actions")
 return ____exports
- end,
-}
-return require("index", ...)
