+++ tsconfig.json
@@ -1,14 +1,7 @@
 {
-  "compilerOptions": {
-    "target": "esnext",
-    "lib": ["esnext"],
-    "moduleResolution": "node",
-    "types": ["@typescript-to-lua/language-extensions", "lua-types/jit"],
-    "strict": true,
-    "esModuleInterop": true,
-    "baseUrl": "src",
-    "paths": {
-      "@ts/*": ["*"]
-    }
-  }
+  "files": [],
+  "references": [
+    { "path": "./tsconfig.lua.json" },
+    { "path": "./tsconfig.node.json" }
+  ]
 }
